requirements.txt
django==4.0.3
django-tastypie
gunicorn
whitenoise